Ogre/write flac

+++ Makefile
@@ -2,8 +2,14 @@ SOURCE = opengate
 RDIR = target/release/
 OPENGATE = $(RDIR)opengate
 
+install-ubuntu-deps:
+	sudo apt install libflac-dev
+
 build:
 	cargo build --release
+
+install: build
+	cargo install --path .
 
 lint:
 	cargo fmt
@@ -12,7 +18,16 @@ lint:
 test:
 	cargo test --verbose
 
+clean:
+	cargo clean
+
 play-short: build
-	rm ./test_short.wav
-	$(OPENGATE) ./beats/test_short.yaml
+	test -f "./test_short.wav" && rm ./test_short.wav || true
+	$(OPENGATE) ./beats/test_short.yaml --out ./test_short.wav
 	aplay ./test_short.wav
+
+play-short-flac: build
+	test -f "./test_short.flac" && rm ./test_short.flac || true
+	$(OPENGATE) ./beats/test_short.yaml --out ./test_short.flac
+	# Need sudo apt install ffmpeg for ffplay
+	ffplay -autoexit -nodisp ./test_short.flac
